@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
-
+using Microsoft.Extensions.Hosting;
 using TextMood.Shared;
 
 namespace TextMood.SignalR
@@ -10,12 +10,13 @@ namespace TextMood.SignalR
     {
         public void ConfigureServices(IServiceCollection services) => services.AddSignalR();
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
                 app.UseDeveloperExceptionPage();
 
-            app.UseSignalR(route => route.MapHub<TextMoodModelHub>($"/{SignalRConstants.TextMoodModelHubName}"));
+            app.UseRouting();
+            app.UseEndpoints(route => route.MapHub<TextMoodModelHub>($"/{SignalRConstants.TextMoodModelHubName}"));
         }
     }
 }
